Tidy up of a bunch of code
