+++ .gitattributes
@@ -1,0 +1,6 @@
+*fits binary
+*fit binary
+*fts binary
+*fit.gz binary
+*fits.gz binary
+*fts.gz binary
