+++ .github/workflows/ci.yml
@@ -1,0 +1,122 @@
+name: CI
+
+on:
+  push:
+    branches:
+      - 'main'
+      - '*.*'
+      - '!*backport*'
+    tags:
+      - 'v*'
+      - '!*dev*'
+      - '!*pre*'
+      - '!*post*'
+  pull_request:
+  workflow_dispatch:
+
+concurrency:
+  group: ${{ github.workflow }}-${{ github.ref }}
+  cancel-in-progress: true
+
+jobs:
+  core:
+    uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@main
+    with:
+      submodules: false
+      coverage: codecov
+      posargs: -n auto
+      libraries: |
+        apt:
+          - ffmpeg
+      envs: |
+        - linux: py310
+
+  test:
+    needs: [core]
+    uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@main
+    with:
+      submodules: false
+      coverage: codecov
+      posargs: -n auto
+      libraries: |
+        apt:
+          - ffmpeg
+        brew:
+          - ffmpeg
+      envs: |
+        - linux: py311
+        - windows: py39
+        - macos: py38
+        - linux: py38-oldestdeps
+
+  docs:
+    needs: [core]
+    uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@main
+    with:
+      default_python: '3.8'
+      submodules: false
+      pytest: false
+      cache-key: docs-${{ github.run_id }}
+      libraries: |
+        apt:
+          - ffmpeg
+          - graphviz
+      envs: |
+        - linux: build_docs
+
+  online:
+    if: "!startsWith(github.event.ref, 'refs/tags/v')"
+    needs: [core]
+    uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@main
+    with:
+      default_python: '3.8'
+      submodules: false
+      coverage: codecov
+      posargs: -n auto --dist loadgroup
+      libraries: |
+        apt:
+          - ffmpeg
+      envs: |
+        - linux: py39-online
+
+  cron:
+    if: |
+      github.event_name == 'workflow_dispatch' || (
+        github.event_name == 'pull_request' &&
+        contains(github.event.pull_request.labels.*.name, 'Run cron CI')
+      )
+    uses: OpenAstronomy/github-actions-workflows/.github/workflows/tox.yml@main
+    with:
+      default_python: '3.8'
+      submodules: false
+      coverage: codecov
+      posargs: -n auto
+      libraries: |
+        apt:
+          - ffmpeg
+      envs: |
+        - linux: py311-devdeps
+        - linux: py39-conda
+          libraries: ''
+
+  publish:
+    # Build wheels when pushing to any branch except main
+    # publish.yml will only publish if tagged ^v.*
+    if: |
+      (
+        github.event_name != 'pull_request' && (
+          github.ref_name != 'main' ||
+          github.event_name == 'workflow_dispatch'
+        )
+      ) || (
+        github.event_name == 'pull_request' &&
+        contains(github.event.pull_request.labels.*.name, 'Run publish')
+      )
+    needs: [test]
+    uses: OpenAstronomy/github-actions-workflows/.github/workflows/publish_pure_python.yml@main
+    with:
+      test_extras: 'dev'
+      test_command: 'pytest -p no:warnings --doctest-rst -m "not mpl_image_compare" --pyargs irispy'
+      submodules: false
+    secrets:
+      pypi_token: ${{ secrets.pypi_token }}
