+++ .gitignore
@@ -220,11 +220,6 @@ irispy/_version.py
 .history
 *.orig
 .tmp
+
+# Example files
 iris_l2_20140708_114109_3824262996_raster_t000_r00000.fits
-iris_l2_20130902_163935_4000255147_SJI_1400_t000.fits.gz
-iris_l2_20130902_163935_4000255147_raster.tar.gz
-iris_l2_20130902_163935_4000255147_raster_t000_r00000.fits
-iris_l2_20140708_114109_3824262996_raster.tar.gz
-iris_l2_20140919_051712_3860608353_SJI_2832_t000.fits.gz
-iris_l2_20180102_153155_3610108077_raster.tar.gz
-iris_l2_20180102_153155_3610108077_raster_t000_r00000.fits
