--- .gitlab-ci.yml
@@ -1,68 +1,0 @@
-image: python:3.10
-
-before_script:
-  - apt update -qq
-  - apt install -qq --yes ffmpeg
-  - pip install tox
-
-stages:
-  - Basic
-  - Test
-  - Extra
-  - Deploy
-
-codestyle:
-  stage: Basic
-  script:
-    - tox -e codestyle
-
-py310:
-  stage: Basic
-  script:
-    - tox -e py310
-
-py39:
-  stage: Test
-  image: python:3.9
-  script:
-    - tox -e py39
-
-py38-online:
-  stage: Test
-  image: python:3.8
-  script:
-    - tox -e py38-online
-    - pip install --upgrade codecov
-    - codecov
-
-py38-devdeps:
-  allow_failure: true
-  stage: Extra
-  image: python:3.8
-  script:
-    - tox -e py38-devdeps
-
-build_docs:
-  stage: Extra
-  before_script:
-    # Need graphviz binary for generating inheritance diagrams
-    - apt-get update -qq && apt-get install -y -qq graphviz
-    - pip install tox
-  script:
-    - tox -e build_docs
-
-# Upload to PyPI when we push to a tag
-pypi:
-  stage: Deploy
-  # only run on parent tags with semantic versioning names, excluding dev tags
-  only:
-      - tags
-  except:
-    - branches
-    - merge_requests
-  before_script:
-    - python -m pip install --upgrade build check-manifest twine
-  script:
-    - python -m build --sdist --wheel --outdir dist/ .
-    - twine check dist/*
-    - twine upload dist/irispy-lmsal*
