+++ .readthedocs.yml
@@ -1,20 +1,21 @@
-# Read the Docs configuration file
-# See https://docs.readthedocs.io/en/stable/config-file/v2.html for details
 version: 2
 build:
-  os: ubuntu-20.04
+  os: ubuntu-22.04
   tools:
-    python: "3.9"
-  apt_packages:
-    - graphviz
-
+    python: "mambaforge-4.10"
+  jobs:
+    pre_install:
+      - git update-index --assume-unchanged .rtd-environment.yml docs/conf.py
+conda:
+  environment: .rtd-environment.yml
 sphinx:
+  builder: html
   configuration: docs/conf.py
-formats: []
+  fail_on_warning: false
 python:
-  install:
-    - method: pip
-      path: .
-      extra_requirements:
-        - docs
+   install:
+     - method: pip
+       extra_requirements:
         - all
+        - docs
+       path: .
