+++ .rtd-environment.yml
@@ -1,0 +1,7 @@
+name: rtd_irispy
+channels:
+  - conda-forge
+dependencies:
+  - python=3.11
+  - pip
+  - graphviz!=2.42.*,!=2.43.*
