--- MANIFEST.in
@@ -1,17 +1,0 @@
-include README.rst
-include CHANGES.rst
-include setup.cfg
-include LICENSE.rst
-include pyproject.toml
-
-recursive-include irispy *.pyx *.c *.pxd
-recursive-include docs *
-recursive-include licenses *
-recursive-include cextern *
-recursive-include scripts *
-
-prune build
-prune docs/_build
-prune docs/api
-
-global-exclude *.pyc *.o
