+++ README.rst
@@ -1,3 +1,4 @@
+****************
 ``irispy-lmsal``
 ****************
 
@@ -5,8 +6,7 @@
   :width: 400
   :alt: Image of the IRIS spacecraft
 
-``irispy-lmsal`` is a library that provides the tools to read in and analyze data from Interface Region
-Imaging Spectrograph (IRIS).
+``irispy-lmsal`` is a library that provides the tools to read in and analyze data from Interface Region Imaging Spectrograph (IRIS).
 
 IRIS is a NASA-funded Small Explorer which uses a high-frame-rate ultraviolet imaging spectrometer to make observations of the Sun.
 For more information see the mission/instrument paper which is available `online for free <https://www.lmsal.com/iris_science/doc?cmd=dcur&proj_num=IS0196&file_type=pdf>`__.
