+++ docs/code_api/iripsy.rst
@@ -1,0 +1,7 @@
+*********************
+``irispy`` (`irispy`)
+*********************
+
+The `irispy` library provides tools to read and work with IRIS data.
+
+.. automodapi:: irispy
