+++ docs/code_api/sji.rst
@@ -1,6 +1,7 @@
+******************
 SJI (`irispy.sji`)
 ******************
 
-The ``sji`` module provides tools to read and work with IRIS Level 2 SJI data.
+The `irispy.sji` module provides tools to read and work with level 2 SJI data.
 
 .. automodapi:: irispy.sji
