+++ irispy/data/_sample.py
@@ -65,9 +65,7 @@ def _handle_final_errors(results):
     for err in results.errors:
         file_name = err.url.split("/")[-1]
         log.debug(f"Failed to download {_SAMPLE_FILES[file_name]} from {err.url}: {err.exception}")
-        log.error(
-            f"Failed to download {_SAMPLE_FILES[file_name]} from all mirrors," "the file will not be available."
-        )
+        log.error(f"Failed to download {_SAMPLE_FILES[file_name]} from all mirrors," "the file will not be available.")
 
 
 def download_sample_data(overwrite=False):
