+++ irispy/tests/test_sji.py
@@ -1,30 +1,5 @@
 import astropy.units as u
 import numpy as np
-import pytest
-
-from irispy.data.test import get_test_filepath
-from irispy.io.sji import read_sji_lvl2
-
-
-@pytest.fixture
-def IRISMapCube_1330():
-    return read_sji_lvl2(get_test_filepath("iris_l2_20170502_052551_3893010094_SJI_1330_t000.fits"))
-
-
-@pytest.fixture
-def IRISMapCube_1400():
-    return read_sji_lvl2(get_test_filepath("iris_l2_20170502_052551_3893010094_SJI_1400_t000.fits"))
-
-
-@pytest.fixture
-def IRISMapCube_2796():
-    return read_sji_lvl2(get_test_filepath("iris_l2_20170502_052551_3893010094_SJI_2796_t000.fits"))
-
-
-@pytest.fixture
-def IRISMapCube_2832():
-    return read_sji_lvl2(get_test_filepath("iris_l2_20170502_052551_3893010094_SJI_2832_t000.fits"))
-
 
 AXIS = [
     ("meta.obs.sequence",),
@@ -34,20 +9,20 @@ AXIS = [
 
 
 def test_world_axis_physical_types_IRISMapCube_2832(IRISMapCube_2832):
-    assert np.all(IRISMapCube_2832.dimensions == [16, 77, 148] * u.pix)
+    assert np.all(IRISMapCube_2832.dimensions == [10, 40, 37] * u.pix)
     assert IRISMapCube_2832.array_axis_physical_types == AXIS
 
 
 def test_world_axis_physical_types_IRISMapCube_2796(IRISMapCube_2796):
-    assert np.all(IRISMapCube_2796.dimensions == [16, 77, 148] * u.pix)
+    assert np.all(IRISMapCube_2796.dimensions == [62, 40, 37] * u.pix)
     assert IRISMapCube_2796.array_axis_physical_types == AXIS
 
 
 def test_world_axis_physical_types_IRISMapCube_1400(IRISMapCube_1400):
-    assert np.all(IRISMapCube_1400.dimensions == [16, 77, 148] * u.pix)
+    assert np.all(IRISMapCube_1400.dimensions == [62, 40, 37] * u.pix)
     assert IRISMapCube_1400.array_axis_physical_types == AXIS
 
 
 def test_world_axis_physical_types_IRISMapCube_1330(IRISMapCube_1330):
-    assert np.all(IRISMapCube_1330.dimensions == [16, 77, 148] * u.pix)
+    assert np.all(IRISMapCube_1330.dimensions == [52, 40, 37] * u.pix)
     assert IRISMapCube_1330.array_axis_physical_types == AXIS
