+++ irispy/version.py
@@ -6,9 +6,7 @@ try:
 except Exception:
     import warnings
 
-    warnings.warn(
-        f'could not determine {__name__.split(".")[0]} package version; this indicates a broken installation'
-    )
+    warnings.warn(f'could not determine {__name__.split(".")[0]} package version; this indicates a broken installation')
     del warnings
 
     version = "0.0.0"
