+++ setup.cfg
@@ -1,16 +1,23 @@
 [metadata]
 name = irispy-lmsal
+provides = irispy
 author = IRIS Instrument Team
 author_email = freij@baeri.org
 license = BSD 3-Clause
-license_file = LICENSE.rst
-url = https://gitlab.com/LMSAL_HUB/iris_hub/irispy-lmsal
-description = Provides tools to read in and analyze data from the IRIS solar-observing satellite.
+license_files = LICENSE.rst
+url = https://github.com/LM-SAL/irispy-lmsal
+download_url = https://pypi.org/project/irispy-lmsal/
+description = Provides tools to read and analyze data from the IRIS solar-observing satellite.
 long_description = file: README.rst
 long_description_content_type = text/x-rst
-gitlab_project = LMSAL_HUB/iris_hub/irispy-lmsal
+github_project = LMSAL_HUB/iris_hub/irispy-lmsal
 platform = any
-keywords = solar physics, solar, science, sun, wcs, coordinates
+keywords = solar physics, solar, science, sun, wcs, iris, lmsal
+project_urls=
+  Source Code = https://github.com/LM-SAL/irispy-lmsal
+  Documentation = https://irispy-lmsal.readthedocs.io/en/stable/
+  Changelog = https://irispy-lmsal.readthedocs.io/en/stable/changelog.html
+  Issue Tracker = https://github.com/LM-SAL/irispy-lmsal/issues
 classifiers =
     Development Status :: 3 - Alpha
     Intended Audience :: Science/Research
@@ -22,6 +29,7 @@ classifiers =
     Programming Language :: Python :: 3.8
     Programming Language :: Python :: 3.9
     Programming Language :: Python :: 3.10
+    Programming Language :: Python :: 3.11
     Topic :: Scientific/Engineering :: Physics
 
 [options]
@@ -31,7 +39,7 @@ python_requires = >=3.8
 setup_requires = setuptools_scm
 install_requires =
     sunraster>=0.4.0
-    sunpy[all]>=3.1.0
+    sunpy[all]>=4.0.0
     scipy>=1.7.0
     pandas>=1.3.3
 
@@ -40,7 +48,7 @@ tests =
     pytest-astropy
 docs =
     aiapy
-    requests
+    pooch
     sphinx
     sphinx-automodapi
     sphinx-changelog
@@ -72,6 +80,8 @@ filterwarnings =
     #
     # This is due to dependencies building with a numpy version different from
     # the local installed numpy version, but should be fine
+      # https://github.com/pytest-dev/pytest-cov/issues/557
+    ignore:The --rsyncdir command line argument and rsyncdirs config variable are deprecated.:DeprecationWarning
     # See https://github.com/numpy/numpy/issues/15748#issuecomment-598584838
     ignore:The distutils.sysconfig module is deprecated, use sysconfig instead
     ignore:numpy.ufunc size chaEged:RuntimeWarning
@@ -139,3 +149,8 @@ exclude_lines =
   pragma: py{ignore_python_version}
   # Don't complain about IPython completion helper
   def _ipython_key_completions_
+
+[codespell]
+skip = *.asdf,*.fits,*.fts,*.header,*.json,*.xsh,*cache*,*egg*,*extern*,.git,.idea,.tox,_build,*truncated
+ignore-words-list =
+  sav
