+++ setup.py
@@ -3,7 +3,12 @@ from setuptools import setup  # isort:skip
 import os
 from itertools import chain
 
-from setuptools.config import read_configuration
+try:
+    # Recommended for setuptools 61.0.0+
+    # (though may disappear in the future)
+    from setuptools.config.setupcfg import read_configuration
+except ImportError:
+    from setuptools.config import read_configuration
 
 ################################################################################
 # Programmatically generate some extras combos.
