+++ tox.ini
@@ -1,27 +1,15 @@
 [tox]
 envlist =
-    py{38,39,310}{,-oldestdeps,-devdeps,-online}
+    py{38,39,310,311}{,-oldestdeps,-devdeps,-online}
     build_docs
     codestyle
 isolated_build = true
 
 [testenv]
-pypi_filter = file://.test_package_pins.txt
-# We use bash in some of our enviroments so we have to whitelist it.
 whitelist_externals=
     /bin/bash
     /usr/bin/bash
-# Run the tests in a temporary directory to make sure that we don't import
-# irispy from the source tree
 changedir = .tmp/{envname}
-# tox environments are constructued with so-called 'factors' (or terms)
-# separated by hyphens, e.g. test-devdeps-cov. Lines below starting with factor:
-# will only take effect if that factor is included in the environment name. To
-# see a list of example environments that can be run, along with a description,
-# run:
-#
-#     tox -l -v
-#
 description =
     run tests
     devdeps: with the latest developer version of key dependencies
@@ -33,22 +21,18 @@ setenv =
     PYTEST_COMMAND = pytest -vvv -s -raR --pyargs irispy --cov-report=xml --cov=irispy --cov-config={toxinidir}/setup.cfg {toxinidir}/docs
     py37: PYTEST_COMMAND = pytest -vvv -s -raR --pyargs irispy -p no:warnings --cov-report=xml --cov=irispy --cov-config={toxinidir}/setup.cfg {toxinidir}/docs
 deps =
-    # We need this for some packages.
-    setuptools
     # All our tests run in parallel which this plugin allows us to.
     pytest-xdist
     # We check codecov in tox builds.
     pytest-cov
-    # The devdeps factor is intended to be used to install the latest developer version.
-    # of key dependencies.
+    # The devdeps factor is intended to be used to install the latest developer version of key dependencies.
     devdeps: git+https://github.com/sunpy/sunraster
+    devdeps: git+https://github.com/sunpy/ndcube
     # Oldest deps we pin against.
-    oldestdeps: sunraster<0.3
+    oldestdeps: sunraster<0.5
     # These are specific online extras we use to run the online tests.
     online: pytest-rerunfailures
     online: pytest-timeout
-# The following indicates which extras_require from setup.cfg will be installed
-# dev is special in that it installs everything
 extras =
     dev
 commands =
@@ -64,11 +48,10 @@ commands =
     python -c 'import pathlib; print("Documentation available under file://\{0\}".format(pathlib.Path(r"{toxinidir}") / "docs" / "_build" / "index.html"))'
 
 [testenv:codestyle]
-pypi_filter =
 skip_install = true
 description = Run all style and file checks with pre-commit
 deps =
     pre-commit
 commands =
     pre-commit install-hooks
-    pre-commit run --all-files
+    pre-commit run --color always --all-files --show-diff-on-failure
